General improvements to query string handling
Deprecated the addressable gem in favor of HTTPClient's URIBuilder. Yay dependency removal!

Additionally, Manticore will now treat `:params` as equivalent to `:query` for GET, HEAD, and DELETE requests,
in the cases that `:query` is not passed. If you pass both `:query` and `:params` to one of these verbs for
some reason, then :query will be added to the query string, while `:params` will be passed as a request body
(or ignored, depending on the HTTP verb).

Added examples and documentation for :query vs :params

Improved Ruby Hash => request body building. It should now be possible to pass deeply-nested hashes containing
Arrays, Hashes, and non-String leaf values such as Symbols or Fixnums and they'll get properly converted for
the purposes of building the query string or request body.