BREAKING CHANGE: Unify background request semantics with sync/async semantics
Background requests (`client.background.get(...)`) now return a Manticore::Response
rather than immediately passing the request to the executor and returning a Future.

This allows users to construct background requests which can be fired off in a non-blocking
manner with the same interface and usage mechanics as other request types. The primary
motivation for this is to allow for attaching on_success handlers and the like in the same
way that you attach them for other types of requests.

BEFORE:

    retval = client.background.get(...) # => Request is running, returns a Future

AFTER:

    retval = client.background.get(...) # => Request hasn't run yet, returns a Response
    retval.call # => Queues the Response in the Executor and returns a Future

To upgrade your prior background usage to be compatible with Manticore 0.5+, you should invoke #call
on the results of your background requests (ie, `client.background.get(...)` becomes
`client.background.get(...).call`).

Additionally, in an attempt to clarify `async` vs `background` usage, `async` now has two
aliases, `batch` and `parallel`. The following are equivalent:

    client.async.get(...)
    client.batch.get(...)
    client.parallel.get(...)
    client.execute! # All 3 queued requests invoked. Calling thread blocks til all requests complete.