Add support for repeated headers. Fixes #57
If a client responds with multiple headers, Manticore would previously overwrite
earlier values with later values, in violation of RFC 7230. Manticore will now
aggregate headers into a hash of {key => value(s)}, where value(s) may be a
primitive type (in the common case) or an array of values (in the repeated case).
For example, if a response responds with:

   Content-Type: text/plain
   Content-Length: 1500
   X-Header: 1
   X-Header: 2

Then the resulting response#headers hash will be:

   {
     "Content-Type" => "text/plain",
     "Content-Length" => 1500,
     "X-Header" => [1, 2]
   }

This does necessitate that users check the type of a header value prior to using it
if a header may respond with multiple values. To cover both cases, one might consume
headers via something like:

   Array(response["Link"]).each {|value| ... }

Or, if you don't care about headers beyond the first:

   Array(response["Link"]).first

In the future, Manticore may move to returning arrays for even single-value headers,
but this change was made to maintain as much backwards compatibility as possible while
allowing for repeated headers.