ssl[:trust_strategy]: wrapper for org.apache.http.conn.ssl.TrustStrategy
Adds an advanced feature so consumers of this library can provide their own
TrustStrategy to refine the situations where the TrustManager can be bypassed.

Consumers can pass either native java implementations of TrustStrategy, or a
ruby-defined Proc that is expected to handle ruby-OpenSSL X509 Certificates.

Works with the trust_strategy that is possibly instantiated with some
values of `ssl[:verify]` to produce a _combined_ TrustStrategy that will
bypass the TrustManager if _any_ of the configured trust strategies indicate
that the cert chain qualifies for a bypass.