don't send entity when get/delete have no body
Currently a non-body get/delete will sent a "Content-Length: 0"
header given the httpclient is passed a request object with Entity.

While RFC 7230 does allow for non-body requests to have
"Content-Length: 0" it is marked as a SHOULD NOT and can cause problems
in stricter environments.

This commit changes that, where a no body request will use non-entity
objects.